move dd wrapped provider to rpc adapter pkg
