+++ packages/eth-providers/package.json
@@ -17,7 +17,6 @@
     "@acala-network/contracts": "4.3.4",
     "@acala-network/eth-transactions": "workspace:*",
     "bn.js": "~5.2.0",
-    "dd-trace": "~4.16.0",
     "ethers": "~5.7.0",
     "graphql": "~16.0.1",
     "graphql-request": "~3.6.1",
@@ -26,7 +25,6 @@
   "devDependencies": {
     "@acala-network/api": "~6.0.4",
     "@types/bn.js": "~5.1.0",
-    "@types/dd-trace": "^0.9.0",
     "@types/lru-cache": "~7.6.1",
     "dotenv": "~10.0.0",
     "jsdom": "^22.0.0",
